separated progress bar component, added conditional display of plural quantity
